Trim dependencies
