+++ test/index.spec.js
@@ -1,17 +1,16 @@
 const { expect } = require('chai');
-const React = require('react');
-const ExecutionEnvironment = require('exenv');
+const { h, Component, render } = require('preact');
+/** @jsx h */
 const jsdom = require('jsdom');
-const { shallow, mount } = require('enzyme')
-const { renderToStaticMarkup } = require('react-dom/server')
-const { render } = require('react-dom')
+const renderToString = require('preact-render-to-string').render;
+const shallow = require('preact-render-to-string').shallowRender;
 
 const withSideEffect = require('../src');
 
 function noop() { }
 const identity = x => x
 
-describe('preact-side-effect (on React)' , () => {
+describe('preact-side-effect', () => {
   describe('argument validation', () => {
     it('should throw if no reducePropsState function is provided', () => {
       expect(withSideEffect).to.throw('Expected reducePropsToState to be a function.');
@@ -33,15 +32,8 @@ describe('preact-side-effect (on React)' , () => {
   describe('displayName', () => {
     const withNoopSideEffect = withSideEffect(noop, noop);
 
-    it('should wrap the displayName of wrapped createClass component', () => {
-      const DummyComponent = React.createClass({displayName: 'Dummy', render: noop});
-      const SideEffect = withNoopSideEffect(DummyComponent);
-
-      expect(SideEffect.displayName).to.equal('SideEffect(Dummy)');
-    });
-
     it('should wrap the displayName of wrapped ES2015 class component', () => {
-      class DummyComponent extends React.Component {
+      class DummyComponent extends Component {
         static displayName = 'Dummy'
         render() {}
       }
@@ -59,15 +51,14 @@ describe('preact-side-effect (on React)' , () => {
     });
 
     it('should fallback to "Component"', () => {
-      const DummyComponent = React.createClass({displayName: null, render: noop});
-      const SideEffect = withNoopSideEffect(DummyComponent);
+      const SideEffect = withNoopSideEffect(function(){});
 
       expect(SideEffect.displayName).to.equal('SideEffect(Component)');
     });
   });
 
   describe('SideEffect component', () => {
-    class DummyComponent extends React.Component {
+    class DummyComponent extends Component {
       render () {
         return <div>hello {this.props.foo}</div>
       }
@@ -81,7 +72,7 @@ describe('preact-side-effect (on React)' , () => {
     });
 
     it('should expose the canUseDOM flag', () => {
-      expect(SideEffect).to.have.property('canUseDOM', ExecutionEnvironment.canUseDOM);
+      expect(SideEffect).to.have.property('canUseDOM', false);
     });
 
     describe('rewind', () => {
@@ -172,7 +163,7 @@ describe('preact-side-effect (on React)' , () => {
     });
 
     it('should render the wrapped component', () => {
-      const markup = renderToStaticMarkup(<SideEffect foo="bar"/>);
+      const markup = renderToString(<SideEffect foo="bar"/>);
 
       expect(markup).to.equal('<div>hello bar</div>');
     });
@@ -180,12 +171,19 @@ describe('preact-side-effect (on React)' , () => {
     describe('with DOM', () => {
       const originalWindow = global.window;
       const originalDocument = global.document;
+      const originalSVGElement = global.SVGElement;
+      const originalText = global.Text;
 
       before(done => {
         jsdom.env('<!doctype html><html><head></head><body></body></html>', (error, window) => {
           if (!error) {
             global.window = window;
             global.document = window.document;
+            // preact tests for SVGElement, but jsdom does not support it
+            // https://github.com/tmpvar/jsdom/issues/1423
+            global.SVGElement = function(){};
+            // Preact tests for Text, supported by jsdom
+            global.Text = window.Text
           }
 
           done(error);
@@ -195,17 +193,8 @@ describe('preact-side-effect (on React)' , () => {
       after(() => {
         global.window = originalWindow;
         global.document = originalDocument;
-      });
-
-      it('should be findable by react TestUtils', () => {
-        class AnyComponent extends React.Component {
-          render() {
-            return <SideEffect foo="bar" />
-          }
-        }
-        const wrapper = shallow(<AnyComponent />);
-        const sideEffect = wrapper.find(SideEffect)
-        expect(sideEffect.props()).to.deep.equal({ foo: 'bar' });
+        global.SVGElement = originalSVGElement;
+        global.Text = originalText;
       });
 
       it('should only recompute when component updates', () => {
@@ -222,13 +211,18 @@ describe('preact-side-effect (on React)' , () => {
         const node = document.createElement('div');
         document.body.appendChild(node);
 
-        render(<SideEffect text="bar" />, node);
+        // preact diffs against actual DOM. 
+        // Pass node.firstChild as 3rd argument to render
+        render(<SideEffect text="bar" />, node, node.firstChild);
         expect(collectCount).to.equal(1);
-        render(<SideEffect text="bar" />, node);
+
+        render(<SideEffect text="bar" />, node, node.firstChild);
         expect(collectCount).to.equal(1);
-        render(<SideEffect text="baz" />, node);
+
+        render(<SideEffect text="baz" />, node, node.firstChild);
         expect(collectCount).to.equal(2);
-        render(<SideEffect text="baz" />, node);
+
+        render(<SideEffect text="baz" />, node, node.firstChild);
         expect(collectCount).to.equal(2);
       });
     });
